# backend/api-docs/swagger.yaml
openapi: 3.0.3
info:
  title: FES Stories API
  description: |
    API documentation for FES Stories application.
    
    This API provides endpoints for authentication, health checks, user management, activity tracking, story categories, and story management.
    
    **Base URL:** `http://localhost:8000`
    
    **Authentication:** Currently not required for public endpoints (will be added for protected endpoints)
    
    **Activity Tracking:** All user activities are logged to the database. Users can only view their own activities.
    
    **Story Management:** 
    - Readers can submit stories for review
    - Super admins can review, approve, reject, edit, and delete stories
    - Published stories are available via public endpoint
  version: 1.0.0
  contact:
    name: FES Stories API Support
    email: support@fesstories.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000
    description: Local development server
  - url: https://api.fesstories.com
    description: Production server (when available)
...
paths:
...
  /api/users/roles:
    get:
      tags:
        - Users
      summary: Get Available Roles
      description: |
        Get all available roles excluding Super admin.
        
        **Access:** Super admin only
        
        Returns a list of roles that can be assigned to users.
      operationId: getRoles
      responses:
        '200':
          description: Roles retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoleListResponse'
              example:
                success: true
                roles:
                  - id: 2
                    role_name: "Reader"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                message: "An error occurred while fetching roles"
...
components:
  schemas:
...
    RoleItem:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Role ID
          example: 2
        role_name:
          type: string
          description: Role name
          example: "Reader"
    RoleListResponse:
      type: object
      required:
        - success
        - roles
      properties:
        success:
          type: boolean
          example: true
        roles:
          type: array
          items:
            $ref: '#/components/schemas/RoleItem'
    ErrorResponse:
      type: object
      required:
        - success
        - message
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          description: Error message
          example: "Invalid credentials"
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token authentication (to be implemented)